Add a dedicated form rendering route
This makes it easier to add new forms by simply creating the needed files and linking to the form route with the form name.
