+++ tests/dummy/app/components/test-form.hbs
@@ -1,13 +1,9 @@
-{{#if this.loadFormData.isRunning}}
-  <AuLoader @padding="small" />
-{{else}}
-  <RdfForm
-    @groupClass="au-o-grid__item"
-    @form={{this.form}}
-    @graphs={{this.graphs}}
-    @sourceNode={{this.sourceNode}}
-    @formStore={{this.formStore}}
-    @show={{this.formConfig.isReadOnly}}
-    @forceShowErrors={{false}}
-  />
-{{/if}}
+<RdfForm
+  @groupClass="au-o-grid__item"
+  @form={{@form}}
+  @graphs={{this.graphs}}
+  @sourceNode={{this.sourceNode}}
+  @formStore={{@formStore}}
+  @show={{this.formConfig.isReadOnly}}
+  @forceShowErrors={{false}}
+/>
