+++ tests/dummy/app/routes/form.js
@@ -1,0 +1,63 @@
+import Route from '@ember/routing/route';
+import { ForkingStore } from '@lblod/ember-submission-form-fields';
+import rdflib from 'browser-rdflib';
+
+const FORM_GRAPHS = {
+  formGraph: new rdflib.NamedNode('http://data.lblod.info/form'),
+  metaGraph: new rdflib.NamedNode('http://data.lblod.info/metagraph'),
+  sourceGraph: new rdflib.NamedNode(`http://data.lblod.info/sourcegraph`),
+};
+
+const FORM = new rdflib.Namespace('http://lblod.data.gift/vocabularies/forms/');
+const RDF = new rdflib.Namespace('http://www.w3.org/1999/02/22-rdf-syntax-ns#');
+
+const FORM_TITLES = {
+  'basic-fields': 'Basic form fields',
+};
+
+export default class FormRoute extends Route {
+  async model({ formName }) {
+    let [formTtl, metaTtl] = await Promise.all([
+      fetchForm(formName),
+      fetchFormMeta(formName),
+    ]);
+
+    let formStore = new ForkingStore();
+    formStore.parse(formTtl, FORM_GRAPHS.formGraph, 'text/turtle');
+    formStore.parse(metaTtl, FORM_GRAPHS.metaGraph, 'text/turtle');
+
+    let form = formStore.any(
+      undefined,
+      RDF('type'),
+      FORM('Form'),
+      FORM_GRAPHS.formGraph
+    );
+
+    this.form = form;
+    this.formStore = formStore;
+    return {
+      formName,
+      form,
+      formStore,
+      title: FORM_TITLES[formName],
+    };
+  }
+}
+
+async function fetchForm(formName) {
+  let response = await fetch(getFormDataPath(formName, 'form.ttl'));
+  let ttl = await response.text();
+
+  return ttl;
+}
+
+async function fetchFormMeta(formName) {
+  let response = await fetch(getFormDataPath(formName, 'meta.ttl'));
+  let ttl = await response.text();
+
+  return ttl;
+}
+
+function getFormDataPath(formName, fileName) {
+  return `/test-forms/${formName}/${fileName}`;
+}
