+++ tests/dummy/app/templates/application.hbs
@@ -9,7 +9,7 @@
           <nav>
             <ul class="au-c-list-navigation">
               <li class="au-c-list-navigation__item">
-                <AuNavigationLink @route="index">
+                <AuNavigationLink @route="form" @model="basic-fields">
                   Basic form fields
                 </AuNavigationLink>
               </li>
