+++ tests/dummy/app/templates/form.hbs
@@ -1,0 +1,6 @@
+{{page-title @model.title}}
+
+<TestForm
+  @form={{@model.form}}
+  @formStore={{@model.formStore}}
+/>
