--- tests/dummy/app/templates/index.hbs
@@ -1,3 +1,0 @@
-{{page-title "Basic fields"}}
-
-<TestForm @name="basic-fields" />
