+++ tests/dummy/app/templates/loading.hbs
@@ -1,0 +1,1 @@
+<AuLoader @padding="small" />
